Update recipe checksums and versions [2026-04-08]

+++ new_ports/animalcrossing/source/src/actor/ac_house.c
@@ -39,7 +39,7 @@ enum {
 #define aHUS_GET_LIGHT_PERCENT(h) ((h)->arg1_f)
 
 #define aHUS_SET_LOOKS(h, v) ((h)->arg0 = (v))
-#define aHUS_SET_ANIMAL_P(h, v) ((h)->arg1 = (int)(v))
+#define aHUS_SET_ANIMAL_P(h, v) ((h)->arg1 = (intptr_t)(v))
 #define aHUS_SET_ANIMAL_IDX(h, v) ((h)->arg2 = (v))
 #define aHUS_SET_LIGHT_PERCENT(h, v) ((h)->arg1_f = (v))
